Move getgather/browser/chromefleet.py to getgather/browser.py (browser dir no longer needed)
